+++ tutorial/tutorial4_advanced_training.sh
@@ -7,8 +7,15 @@ these fit invocation with the default one.
 "
 
 # Define wherever you want to store results
-DVC_DATA_DPATH=$HOME/data/dvc-repos/toy_data_dvc
-DVC_EXPT_DPATH=$HOME/data/dvc-repos/toy_expt_dvc
+# We will register them with geowatch DVC to abstract away the
+# machine specific paths in the rest of the code.
+# Running this multiple times is idempotent
+geowatch dvc add --path "$HOME/data/dvc-repos/toy_data_dvc" --tag "toy_data_dvc" --name "toy_data_dvc"
+geowatch dvc add --path "$HOME/data/dvc-repos/toy_expt_dvc" --tag "toy_expt_dvc" --name "toy_expt_dvc"
+
+# Now we can access the registered paths as such
+DVC_DATA_DPATH=$(geowatch dvc --tags "toy_data_dvc")
+DVC_EXPT_DPATH=$(geowatch dvc --tags "toy_expt_dvc")
 
 NUM_TOY_TRAIN_VIDS="${NUM_TOY_TRAIN_VIDS:-100}"  # If variable not set or null, use default.
 NUM_TOY_VALI_VIDS="${NUM_TOY_VALI_VIDS:-5}"  # If variable not set or null, use default.
@@ -38,17 +45,15 @@ fi
 
 
 __doc__="
-
 ###############################################
 # DEMO: MultimodalTransformer with LightningCLI
 ###############################################
-
 "
 
 # Training with the baseline MultiModalModel
 DATASET_CODE=ToyDataMSI
 WORKDIR=$DVC_EXPT_DPATH/training/$HOSTNAME/$USER
-EXPERIMENT_NAME=ToyDataMSI_Demo_V002
+EXPERIMENT_NAME=ToyDataMSI_Demo_LightningCLI
 DEFAULT_ROOT_DIR=$WORKDIR/$DATASET_CODE/runs/$EXPERIMENT_NAME
 MAX_STEPS=1000
 TARGET_LR=3e-4
@@ -91,9 +96,9 @@ python -m watch.tasks.fusion fit --config "
       accumulate_grad_batches: 1
       default_root_dir     : $DEFAULT_ROOT_DIR
       accelerator          : gpu
-      devices              : 0,
-      #devices             : 0,1
-      #strategy            : ddp
+      #devices              : 0,
+      devices             : 0,1
+      strategy            : ddp
       check_val_every_n_epoch: 1
       enable_checkpointing: true
       enable_model_summary: true
@@ -103,18 +108,25 @@ python -m watch.tasks.fusion fit --config "
       num_sanity_val_steps: 0
       replace_sampler_ddp: true
       track_grad_norm: 2
+      callbacks:
+          - class_path: pytorch_lightning.callbacks.ModelCheckpoint
+            init_args:
+                monitor: val_loss
+                mode: min
+                save_top_k: 5
+                filename: '{epoch}-{step}-{val_loss:.3f}.ckpt'
+                save_last: true
     initializer:
         init: noop
 "
 
 
 __doc__="
+####################################################
+# DEMO: MultiGPU Training with MultimodalTransformer
+####################################################
 
-#########################
-# DEMO: MultiGPU Training
-#########################
-
-The following command trains a HeterogeneousModel model on two GPUs with DDP
+The following command trains a MultimodalTransformer model on two GPUs with DDP
 
 It seems to be the case that something in our system can cause DDP to hang with
 100% reported GPU utilization (even though it really isn't doing anything).
@@ -124,6 +136,8 @@ References:
     https://github.com/Lightning-AI/lightning/issues/10947
     https://github.com/Lightning-AI/lightning/issues/5319
     https://github.com/Lightning-AI/lightning/discussions/6501#discussioncomment-553152
+    https://discuss.pytorch.org/t/ddp-via-lightning-fabric-training-hang-with-100-gpu-utilization/181046
+    https://discuss.pytorch.org/t/single-machine-ddp-issue-on-a6000-gpu/134869/5
 
 So far we may be able to avoid this if we do some combination of the following:
     * Disable pl_ext.callbacks.BatchPlotter
@@ -134,8 +148,8 @@ So far we may be able to avoid this if we do some combination of the following:
     * Disable pl.callbacks.ModelCheckpoint
 "
 
-DVC_DATA_DPATH=$HOME/data/dvc-repos/toy_data_dvc
-DVC_EXPT_DPATH=$HOME/data/dvc-repos/toy_expt_dvc
+DVC_DATA_DPATH=$(geowatch dvc --tags "toy_data_dvc")
+DVC_EXPT_DPATH=$(geowatch dvc --tags "toy_expt_dvc")
 NUM_TOY_TRAIN_VIDS="${NUM_TOY_TRAIN_VIDS:-100}"  # If variable not set or null, use default.
 NUM_TOY_VALI_VIDS="${NUM_TOY_VALI_VIDS:-5}"  # If variable not set or null, use default.
 TRAIN_FPATH=$DVC_DATA_DPATH/vidshapes_msi_train${NUM_TOY_TRAIN_VIDS}/data.kwcoco.json
@@ -143,7 +157,101 @@ VALI_FPATH=$DVC_DATA_DPATH/vidshapes_msi_vali${NUM_TOY_VALI_VIDS}/data.kwcoco.js
 
 DATASET_CODE=ToyDataMSI
 WORKDIR=$DVC_EXPT_DPATH/training/$HOSTNAME/$USER
-EXPERIMENT_NAME=ToyDataMSI_Demo_V002
+EXPERIMENT_NAME=ToyDataMSI_Demo_MultiModal_DDP
+DEFAULT_ROOT_DIR=$WORKDIR/$DATASET_CODE/runs/$EXPERIMENT_NAME
+MAX_STEPS=10000
+TARGET_LR=3e-4
+WEIGHT_DECAY=$(python -c "print($TARGET_LR * 0.01)")
+CHANNELS="(*):(disparity|gauss,X.2|Y:2:6,B1|B8a,flowx|flowy|distri)"
+
+export CUDA_VISIBLE_DEVICES=0,1
+DDP_WORKAROUND=1 python -m watch.tasks.fusion fit --config "
+seed_everything: 8675309
+data:
+    num_workers          : 2
+    train_dataset        : $TRAIN_FPATH
+    vali_dataset         : $VALI_FPATH
+    channels             : '$CHANNELS'
+    time_steps           : 5
+    chip_dims            : 128
+    batch_size           : 4
+    max_epoch_length     : 1024
+model:
+    class_path: MultimodalTransformer
+    init_args:
+        saliency_weights       : auto
+        name                   : $EXPERIMENT_NAME
+        class_weights          : auto
+        tokenizer              : linconv
+        arch_name              : smt_it_stm_p8
+        decoder                : mlp
+        positive_change_weight : 1
+        negative_change_weight : 0.01
+        stream_channels        : 16
+        class_loss             : 'dicefocal'
+        saliency_loss          : 'focal'
+        saliency_head_hidden   : 5
+        change_head_hidden     : 5
+        global_change_weight   : 0.00
+        global_class_weight    : 0.00
+        global_saliency_weight : 1.00
+        perterb_scale          : 0.001
+optimizer:
+    class_path: torch.optim.AdamW
+    init_args:
+        lr           : $TARGET_LR
+        weight_decay : $WEIGHT_DECAY
+lr_scheduler:
+  class_path: torch.optim.lr_scheduler.OneCycleLR
+  init_args:
+    max_lr          : $TARGET_LR
+    total_steps     : $MAX_STEPS
+    anneal_strategy : cos
+    pct_start       : 0.05
+trainer:
+    accumulate_grad_batches: 24
+    default_root_dir     : $DEFAULT_ROOT_DIR
+    accelerator          : gpu
+    #devices              : 0,
+    devices              : 0,1
+    strategy             : ddp
+    limit_val_batches    : 256
+    limit_train_batches  : 2048
+    num_sanity_val_steps : 0
+    max_epochs           : 360
+    callbacks:
+        - class_path: pytorch_lightning.callbacks.ModelCheckpoint
+          init_args:
+              monitor: val_loss
+              mode: min
+              save_top_k: 5
+              filename: '{epoch}-{step}-{val_loss:.3f}.ckpt'
+              save_last: true
+
+torch_globals:
+    float32_matmul_precision: auto
+"
+
+
+__doc__="
+
+#########################
+# DEMO: MultiGPU Training
+#########################
+
+The following command trains a HeterogeneousModel model on two GPUs with DDP
+"
+
+DVC_DATA_DPATH=$(geowatch dvc --tags "toy_data_dvc")
+DVC_EXPT_DPATH=$(geowatch dvc --tags "toy_expt_dvc")
+NUM_TOY_TRAIN_VIDS="${NUM_TOY_TRAIN_VIDS:-100}"  # If variable not set or null, use default.
+NUM_TOY_VALI_VIDS="${NUM_TOY_VALI_VIDS:-5}"  # If variable not set or null, use default.
+TRAIN_FPATH=$DVC_DATA_DPATH/vidshapes_msi_train${NUM_TOY_TRAIN_VIDS}/data.kwcoco.json
+VALI_FPATH=$DVC_DATA_DPATH/vidshapes_msi_vali${NUM_TOY_VALI_VIDS}/data.kwcoco.json
+
+DATASET_CODE=ToyDataMSI
+WORKDIR=$DVC_EXPT_DPATH/training/$HOSTNAME/$USER
+EXPERIMENT_NAME=ToyDataMSI_Demo_Heterogeneous_DDP
 DEFAULT_ROOT_DIR=$WORKDIR/$DATASET_CODE/runs/$EXPERIMENT_NAME
 MAX_STEPS=10000
 TARGET_LR=3e-4
@@ -207,7 +315,7 @@ ctrl+C. Then it shows how to restart given the checkpoint that was written.
 # Training with the HeterogeneousModel using a very small backbone
 DATASET_CODE=ToyDataMSI
 WORKDIR=$DVC_EXPT_DPATH/training/$HOSTNAME/$USER
-EXPERIMENT_NAME=ToyDataMSI_Demo_V003
+EXPERIMENT_NAME=ToyDataMSI_Demo_CheckpointRestart
 DEFAULT_ROOT_DIR=$WORKDIR/$DATASET_CODE/runs/$EXPERIMENT_NAME
 # Fresh start
 rm -rf "$DEFAULT_ROOT_DIR"